Refactor environment variable access: move getEnvVariable function to sharedUtils and update import in analyticsService
